captain-console/requirements.txt
asgiref==3.2.7
Django==3.0.6
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
Pillow==7.1.2
django-countries==6.1.2
django-credit-cards==0.4.1
